// frontend/src/api/models/memoryChapterTask.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';

export interface MemoryChapterTask {
  attemptCount: number;
  chapterId: string;
  chapterNum: number;
  createdAt: string;
  memoryJobId: string;
  taskStatus: JobStatus;
  updatedAt: string;
}